+++ src/UndefinedBoxInterface.php
@@ -12,6 +12,9 @@ namespace Contao\ImagineSvg;
 
 use Imagine\Image\BoxInterface;
 
+/**
+ * @deprecated use SvgBox::getType() === SvgBox::TYPE_NONE instead
+ */
 interface UndefinedBoxInterface extends BoxInterface
 {
 }